Refactor discovery and calendar packages to have crisp and fuzzy, #12
