+++ tests/pix_framework/discovery/profile_resource_model_profile_differentiated.py
@@ -1,9 +1,8 @@
 from pathlib import Path
 
-from pix_framework.discovery.resource_calendars import CalendarDiscoveryParams, CalendarType
+from pix_framework.discovery.calendar_discovery_parameters import CalendarDiscoveryParameters, CalendarType
 from pix_framework.discovery.resource_model import ResourceModel, discover_resource_model
-from pix_framework.io.event_log import read_csv_log
-from pix_framework.io.event_log import PROSIMOS_LOG_IDS
+from pix_framework.io.event_log import PROSIMOS_LOG_IDS, read_csv_log
 
 assets_dir = Path(__file__).parent.parent / "assets"
 
@@ -16,7 +15,7 @@ def profile_discover_case_arrival_model_differentiated():
     result = discover_resource_model(
         event_log=log,
         log_ids=log_ids,
-        params=CalendarDiscoveryParams(discovery_type=CalendarType.DIFFERENTIATED_BY_RESOURCE),
+        params=CalendarDiscoveryParameters(discovery_type=CalendarType.DIFFERENTIATED_BY_RESOURCE),
     )
 
     assert type(result) is ResourceModel
