+++ tests/pix_framework/discovery/resource_calendar_and_performance/fuzzy/assets/bpmn_models/LoanApp_simplified.bpmn
