+++ tests/pix_framework/discovery/resource_calendar_and_performance/fuzzy/assets/bpmn_models/fuzzy_calendars_model.bpmn
