+++ tests/pix_framework/discovery/resource_calendar_and_performance/fuzzy/assets/csv_logs/fuzzy_calendars_log.csv.gz
