+++ tests/pix_framework/discovery/resource_calendar_and_performance/fuzzy/assets/csv_logs/loan_SC_HU.csv
