+++ tests/pix_framework/discovery/test_case_arrival.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 import pytest
-from pix_framework.calendar.resource_calendar import RCalendar
+from pix_framework.calendar.crisp_resource_calendar import RCalendar
 from pix_framework.discovery.case_arrival import (
     CaseArrivalModel,
     discover_case_arrival_calendar,
